Import ClangSA.jl source tree
Now that static analysis passes on src/*.c without any false positives,
it's time to integrate ClangSA into our regular CI process. Since it's
a bit specific to the exact version of julia, and for convenience,
we import the source tree into julia proper. `make -C src analyzegc`
runs the analysis. ClangSA also comes with its own test suite, which
is imported, but not hooked up (To be done as part of #28330).